move identity client into a separate file
